+++ CHANGELOG.md
@@ -7,7 +7,8 @@ All notable changes to this project will be documented in this file.
 - Lengthen the silence duration that is being sent
 - Fix concurrent modification of the voice stream
 - Add instruction to system prompt (no markdown anymore)
-- Fixed AI Model not outputting texte correctly
+- Fixed AI Model not outputting text correctly
+- Removed the text and audio functionality, it doesn't work properly with the API
 
 ## [1.2.1] - 2025-06-20
 ### Added